--- source/databricks/calculation_engine/package/settlement_report_job/settlement_report_args.py
@@ -1,30 +1,0 @@
-# Copyright 2020 Energinet DataHub A/S
-#
-# Licensed under the Apache License, Version 2.0 (the "License2");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from dataclasses import dataclass
-from datetime import datetime
-
-from package.settlement_report_job.calculation_type import CalculationType
-
-
-@dataclass
-class SettlementReportArgs:
-    report_id: str
-    period_start: datetime
-    period_end: datetime
-    calculation_type: CalculationType
-    split_report_by_grid_area: bool
-    prevent_large_text_files: bool
-    time_zone: str
-    catalog_name: str
