+++ source/databricks/settlement_report/settlement_report_job/settlement_report.py
@@ -19,14 +19,14 @@ from typing import Callable
 
 from opentelemetry.trace import SpanKind, Status, StatusCode, Span
 
-import package.settlement_report_job.logging_configuration as config
-from package.settlement_report_job.settlement_report_args import SettlementReportArgs
-from package.settlement_report_job.settlement_report_job_args import (
+import settlement_report_job.logging_configuration as config
+from settlement_report_job.settlement_report_args import SettlementReportArgs
+from settlement_report_job.settlement_report_job_args import (
     parse_job_arguments,
     parse_command_line_arguments,
 )
-from package.settlement_report_job.spark_initializor import initialize_spark
-from package.settlement_report_job.time_series_factory import create_time_series
+from settlement_report_job.spark_initializor import initialize_spark
+from settlement_report_job.time_series_factory import create_time_series
 
 
 # The start() method should only have its name updated in correspondence with the
